Removed references from rowIdx and localIdx attributes in SparseMatrixElement and MultidiagonalMatrixElement
The row index and local element index are not stored in row views, so
this is necessary to avoid dangling references bound to temporary
objects.

The columnIndex() member returning non-const reference does not make
sense in MultidiagonalMatrixElement, because the multidiagonal and
tridiagonal formats do not allow changing the column index.