CMakeLists.txt: refactored the -march=native and -mtune=native flags for Release builds
The generator expression is not needed anymore (the flags don't have to
be removed from any target anymore, since the Python bindings are in a
separate project now). Instead, it may be needed to remove these flags
globally, such as when building a distribution package (e.g. AUR) or
when building on a heterogeneous cluster where the login and compute
nodes have different CPU architectures (e.g. Helios).