Removed `__launch_bounda__` from CudaReductionKernel and CudaMultireductionKernel
`__launch_bounda__` does not work with clang++ when it compiles CUDA
code. Even with nvcc it does not seem to be important from the
performance point of view. The result is much simpler and has the same
performance. Note that some of the old logic was not used already (e.g.
blocksdPerMultiprocessor in the CudaReductionKernel).