Segments refactoring: split reduction segments to a separate namespace
Unfortunetaly this resulted in a long and unreadable diff, but the idea
is simple: segments are a data abstraction that keeps track of the
indexing, while segments reduction is a specific algorithm that uses the
segments. Hence, splitting this will lead to cleaner code and I believe
also faster compile times and smaller binaries.

I'm pretty sure that the segments reduction kernels work exactly the
same as before, the only changes were due to interface changes during
the decoupling.

Other notes:

- The CSRDefault, CSRScalar, and other aliases are gone. General code
  using Segments directly has been updated, the SpMV implementation in
  SparseMatrix will be fixed in the following commits.
- The CSR segments don't store the kernel instance anymore and the
  kernels with non-static reduceSegments method (e.g. CSRAdaptive) need
  to be instantiated and initialized. The SparseMatrix implementation
  will be fixed in the following commits.