Updated vectorProduct in SparseMatrix for separate segments reduction kernels
All SpMV tests are now fixed. The vectorProduct method is not tested for
SandboxMatrix, because it does not use segments and does not fit with
the remaining SparseMatrix tests. Note that it was not tested even
before - the cpp file of the test was including
SparseMatrixVectorProductTest_CSRScalar.h instead of
SparseMatrixVectorProductTest_SandboxMatrix.h