Added CMakePresets.json file
This is a feature added in CMake 3.19 that allows to share common build
settings with multiple people. The VSCode extension for CMake supports
presets and shows corresponding buttons in the statusbar.

https://cmake.org/cmake/help/latest/manual/cmake-presets.7.html

The file also contains revised CTest options for running the tests,
which will be used later from the install script and CI.