CMakeLists.txt: add utility targets for each TNL component, remove the corresponding configure options
Configure options are meant to configure the build process, not to
select the build targets. The component targets are collected
per-directory using the BUILDSYSTEM_TARGETS property [1] of the
directory as suggested in [2]. Installable components are labeled
explicitly using the COMPONENT keyword of the install() command [3].

Note that since the TNL_BUILD_DOC option is gone, the request to build
documentation results only in warning rather than error when Doxygen is
not found.

[1] https://cmake.org/cmake/help/latest/prop_dir/BUILDSYSTEM_TARGETS.html
[2] https://discourse.cmake.org/t/mechanisms-for-grouping-targets-labels-or-otherwise/1941
[3] https://cmake.org/cmake/help/latest/command/install.html